Functional update
This program now prints out the information for the preset number of pages, which may be increased manually from 10. The user must enter a search term on the command line as an addition argument, otherwise no search term will be entered and the program will take measures to avoid crashing.